trace: make Timestamp value ever 64bit (version D)
Issue #3657

+++ repos/base/include/spec/arm_v7/trace/timestamp.h
@@ -20,7 +20,7 @@
 
 namespace Genode { namespace Trace {
 
-	typedef uint32_t Timestamp;
+	typedef uint64_t Timestamp;
 
 	inline Timestamp timestamp()
 	{
